Fixing Bug with Arrays
